Decreased header size by chaning most atributes to BYTE unstead of int

+++ header.h
@@ -6,6 +6,8 @@
 
 #pragma once
 
+#include <Windows.h>
+
 /***********************************************************
 * Specific variables
 ************************************************************/
@@ -13,16 +15,16 @@
 typedef struct header Header;
 
 struct header {
-	int sid;					// Sender ID
-	int rid;					// Receiver ID
-	int rid2;					// Used for voteOn() error detection and correction
-	int rid3;					// Used for voteOn() error detection and correction
-	int priority;				// Message's pirority
+	BYTE sid;					// Sender ID
+	BYTE rid;					// Receiver ID
+	BYTE rid2;					// Used for voteOn() error detection and correction
+	BYTE rid3;					// Used for voteOn() error detection and correction
+	BYTE priority;				// Message's pirority
 	int seqNum;					// Message's position in the queue
 	long payloadSize;			// Number of bytes in payload after this header
-	int payloadType;			// mERR, mNONE, mTXT, mAUD, numMsgTypes
-	int encryption;				// ERR, NONE, XOR, VIG, numOfEnc
-	int compression;			// cERR, cNONE, cHUF, cRLE, numCompTypes
+	BYTE payloadType;			// mERR, mNONE, mTXT, mAUD, numMsgTypes
+	BYTE encryption;				// ERR, NONE, XOR, VIG, numOfEnc
+	BYTE compression;			// cERR, cNONE, cHUF, cRLE, numCompTypes
 	long uncompressedLength;	// Uncompressed message length
 	int checksum;				// Checksum for payload error detection
 	bool errorDC;				// Enable or disable error detection and correciton
